Add configuration mechanism
This will load configuration from /etc/duffy.yaml if present, or another
configuration file specified on the command line. The configuration keys
are the same as the click option variables.

In the course, specify the name in click.command so the CLI won't be
called differently in tests than for real, and relax tests processing
help output so we don't have to update them any time we change options.

Signed-off-by: Nils Philippsen <nils@redhat.com>