Straighten out sync vs. async DB access
Previously, I assumed that we could base everything regarding database
engines and sessions off of SQLAlchemy's AsyncEngine. However, I didn't
manage to bind its .sync_engine to a synchronous Session... For the time
being at least, we'll configure and initialize the sync and async
engines and sessions separately.

Additionally, require Python >= 3.7 because we need asyncio.current_task
which isn't available in 3.6 and on Python 3.7, disable the pytest
coverage plugin because some tests have to be skipped there.

Signed-off-by: Nils Philippsen <nils@redhat.com>