Add an interactive shell
The new `shell` subcommand initializes the connection to the database
and runs an interactive shell – IPython if available, otherwise plain
Python – with some useful objects in the local namespace.

This adds `ipython` as an optional dependency as part of the new
`interactive` set of extra dependencies.

Additionally, its corresponding CLI test demonstrates how to override
auto-detected values for CLI options of the type click.Choice – this is
needed so that the unit tests run smoothly regardless of if ipython is
available or not.

Fixes: #96

Signed-off-by: Nils Philippsen <nils@redhat.com>