Add TZDateTime and {Creatable,Retirable}Mixin
These mixins can be used in declarative SQLAlchemy classes and add
the `created_at`, `retired_at` and `active` attributes. Accessing or
setting the latter boils down to checking if `retired_at` is set or not,
or setting it to the current time (or not), respectively.

Using TZDateTime rather than DateTime ensures that datetimes are stored
using the UTC time zone in the database. Setting attributes of this type
to a datetime value without associated time zone will be refused. The
corresponding utcnow() can be used for SQL expressions such as in
server-side defaults.

Signed-off-by: Nils Philippsen <nils@redhat.com>