Make API attribute testing more flexible
Previously, all attributes submitted to an endpoint in POST were
expected to be present in the response with the exception of items
proxied to another controller test class for dependencies.

This lets subclasses set `no_response_attrs` to a list of attribute
names which aren't expected in the response and thus won't be verified.

Signed-off-by: Nils Philippsen <nils@redhat.com>