Fix leaky DB sessions
SQLAlchemy scoped sessions need to know the scope for created concrete
sessions. In web frameworks, this would be the serviced request and I
assumed that asyncio tasks would map that well, but they didn't. This
way, concrete DB sessions used in a request weren't cleaned up properly
after raising an exception which caused the lock on the sqlite database
file to not be released, making subsequent writes fail.

This refactors how sessions are handled: `duffy.database` only provides
the `[a]sync_session_maker()` factories which can be used to create
concrete session objects. On the web app side, this is used in the
`req_db_async_session` dependency, which is required by the controller
functions and takes care of creating a session for each request and
cleaning it up again afterwards.

In the course, initialize the database model in a startup event handler
of the app.

Fixes: #158

Signed-off-by: Nils Philippsen <nils@redhat.com>