Don't expose SSH keys in most tenant API responses
This uses pydantic's SecretStr type which hides a value in normal API
responses. We leave it unmangled in create_tenant() responses (the API
user supplied it in the POST body anyway).

Rename BaseTestController.no_response_attrs to the more generic
no_verify_attrs and use it in order to make tenant controller tests not
fail.

Signed-off-by: Nils Philippsen <nils@redhat.com>