Add generic and Ansible mechanism classes
A mechanism implements provisioning and deprovisioning of nodes, the
Mechanism base class provides the basic structure for concrete
subclasses which e.g. delegate the work to Ansible playbooks. It is
always tied to a ConcreteNodePool instance which delegates provisioning
and deprovisioning nodes to it. It uses Jinja2 to expand template
strings in configuration to reference e.g. node pool names.

AnsibleMechanism runs Ansible playbooks to provision and deprovision
nodes for node pools and needs these new dependencies:

- ansible-runner, ansible-core: Python packages to run Ansible playbooks
- ansible: core and community maintained Ansible collections (including
  OpenNebula modules)
- jmespath: needed for the json_query filter which in turn will be used
  to collect information generated during a playbook run and summarize
  it in a format understood by Duffy

The Ansible playbooks aren't part of this repository, if they use
Ansible modules that have other dependencies (like pyone for
community.general.one_vm to talk to an OpenNebula cluster), they'd have
to be installed manually.

Signed-off-by: Nils Philippsen <nils@redhat.com>