Simplify nodes
This refactors Duffy to be agnostic about properties like whether a node
is physical or virtual, its architecture, the installed OS or even what
chassis it's installed in. All this is replaced with the name of the
node pool and a JSON field to store arbitrary info, both of which will
be used by the respective backend mechanism to manage the life cycle of
nodes.

Fixes: #208

Signed-off-by: Nils Philippsen <nils@redhat.com>