Fix passing variables into playbooks
Previously, the `duffy_in` variable wasn't available for templates in
configuration which caused provisioning tasks to raise an exception –
which in turn referenced the wrong missing key because the example
configuration wasn't adapted to later changes in the (de-)provisioning
framework: In an earlier stage of developing the provisioning/deprovisioning
framework, data was passed in and out using an Ansible extra variable or
fact called `duffy`. This didn't ensure that a result is actually set in
the playbook (rather than the input "falling through"), so this was
changed to the extra variable passed in being called `duffy_in` and the
playbook having to set the fact `duffy_out`.

Addtionally, you could only configure one set of extra variables for
both provisioning and deprovisioning playbooks.

Fixes: #242

Signed-off-by: Nils Philippsen <nils@redhat.com>