Fix mapping deprovisioned nodes to results
Previously, this was all wrong: it attempted to match bare fields of the
Node object with output from the deprovisioning playbook. While this
data is passed into the playbook, it's unlikely that the concrete
backend will repeat them back, much rather it will return data about
deprovisioned nodes as it knows them (and what would have ended up in
data.provision when being provisioned).

Signed-off-by: Nils Philippsen <nils@redhat.com>