Lock critical sections in Celery tasks
The critical piece of fill_single_pool() is determining how many nodes
need to be freshly provisioned into a pool. If two tasks do that at the
same time, the pool will end up "over-provisioned". Protecting this with
a lock causes the runner-up task to take the changes that will be done
by the first task into account (and do nothing).

Protecting expire_sessions() with a lock isn't technically necessary but
will prevent a couple of subsequent deprovisioning tasks from stepping
on each other's toes.

Fixes: #269

Signed-off-by: Nils Philippsen <nils@redhat.com>