Use one global lock when provisioning
Previously, the lock was keyed by the pool name but in the case of
reusable nodes that could be used for more than one pool being
provisioned at the same time, this would (in the best case) make only
one transaction succeed and the others fail. As this block shouldn't
take very long to complete, using the lock to serialize access to the
database is a much better option.

Signed-off-by: Nils Philippsen <nils@redhat.com>