Straighten out package extras
Previously, the app serving the API was the base line for dependencies,
but that pulls in a lot of unneeded packages if you only want to use the
client. This defines `app` as its own extra and adds code to deal better
with cases where a command is used that needs another extra of the
`duffy` package to be installed.

Additionally, add Jinja2 as a package for the legacy extra, which was
missing before.

Signed-off-by: Nils Philippsen <nils@redhat.com>