Retry creating session on DB serialization errors
This is necessary because concurrent API calls would all select on the
same set of node objects in separate transactions which fails for every
call but one. Now, create_session() retries the affected code block a
number of times in the case of such an error before it actually flags an
error.

Fixes: #523

Signed-off-by: Nils Philippsen <nils@redhat.com>