Protect more DB actions against serialization errors
At https://pagure.io/centos-infra/issue/1132, an issue is reported where
sessions seem to be created but later can’t be found in the database. My
working hypothesis is that concurrent writes to the DB cause conflicts,
but this only fails when the respective transactions get committed,
which is outside of the request handler when it’s too late to stop the
response that everything worked fine. Mitigate this by retrying more
potentially failing steps and explicitly committing the transaction so a
failure there prevents a response reporting success.

Also, log more information on errors.

Signed-off-by: Nils Philippsen <nils@redhat.com>