Set allowed version ranges of dependencies tighter
Prior to using Renovate to take care of dependency updates, we used
Dependabot which didn’t submit PRs for major version upgrades of
dependencies. With version specifications of the format "^0.y.z", Poetry
only accepts new versions that bump the least significant digit of the
specification as fulfilling the spec.

Fixes: #866

Signed-off-by: Nils Philippsen <nils@redhat.com>